Changed around Object and added the Container class
The base Object class has been reversed, so that subclassing is the
default operation and creating an instance requires calling the 'new'
member.

Container was added as a subclass of Object that extends both the
subclassing and the instantiation mechanisms to allow for defining a
class as a hierarchy of objects. It is kept separate from Object since
this feature does make it a little more heavy-weight.

The ability to create a constructor function has been removed to keep
things simple for now.